+++ MANIFEST.in
@@ -5,5 +5,5 @@ include requirements.txt
 recursive-exclude * __pycache__
 recursive-exclude * *.py[co]
 
-include pygeth/default_blockchain_password
-include pygeth/genesis.json
+include geth/default_blockchain_password
+include geth/genesis.json
