+++ geth/genesis.json
