+++ geth/utils/encoding.py
