+++ geth/utils/filesystem.py
