+++ tests/running/test_running_dev_chain.py
@@ -1,4 +1,4 @@
-from pygeth.geth import DevGethProcess
+from geth.geth import DevGethProcess
 
 
 def test_with_no_overrides(base_dir):
