+++ tox.ini
@@ -17,4 +17,4 @@ deps =
 [testenv:flake8]
 basepython=python
 deps=flake8
-commands=flake8 {toxinidir}/pygeth
+commands=flake8 {toxinidir}/geth
